Add local-path function & shipped mode
The local path depends on whether you have shipped or not. By default it
is the path relative to the system. When shipped it is relative to:

    executable-location/media-folder/system-name/

where the name of the `media-folder` can be customized.

This is only relevant for media included using the manifest. The
c-libraries will have their own directory